pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "twitch-scheduler"
version = "0.0.0"
dependencies = [
    "pyyaml",
    "twitchapi"
]
requires-python = ">=3.12"
description = "Scheduler to watch Twitch for Drops"
readme = "README.md"

[project.urls]
Repository = "https://github.com/me/spam.git"
